requirements.txt
python-telegram-bot==20.4
python-dotenv==1.1.1
groq==0.4.3
